refactor: move home page article list to own component
